Jenkins fix (#2737)
* css update

* added word break too

* fixes

* filter template logic, css fix

* css useUpdateCache

* reset

* FEATURE/HCMPRE-2551: ui ux fixes

* css update

* toggle and app config screen selection issue fix

* ordering in app features screen fixes

* skipping page issue fix

* app config screen tab issue fix

* ui ux fixes

* css update

* next previous positon change , zoom in zoom out added

* css update

* remvoe cache toast and tab switch warning

* demo issue fixes

* css update

* mild ui fixes

* css update

* added update clean fix

* css fixes

* css update

* fixes

* FEATURE/HCMPRE-1551: loc update table submit button issue fix

* HEADING FFIX

* version number nd heading fix

* localisaton table fix

* FEATURE/HCMPRE-2551: id populator option not populating fixes

* field type order fix

* FEATURE/HCMPRE-2551: help text issue fix, mandatory properties hover text

* css update

* helptext localise issue, css fix

* empty messsage localisation and pop up fixes

* loader

* custom fixes

* tooltip fix, change tab fix

* field reorder fix

* localisation fix

* e.some issue potential error fix

* filtering hidden localisation code

* add pop up field error fix

* localisation api call fix

* FEATURE/HCMPRE-2551: Parallel api call for loc update, Clear loc query cache on unmount

* fix

* build fix

* jenkins

* Dockerfile updated

* docker

* upsert localisation hook change

* f

* fix

---------
